requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioresponses==0.7.8
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.10.1
dj-database-url==3.0.1
Django==5.2.4
django-choices-field==2.3.0
djangorestframework==3.16.0
dnspython==2.7.0
docopt==0.6.2
dotenv==0.9.9
drf-spectacular==0.28.0
fastapi==0.116.1
frozenlist==1.7.0
graphql-core==3.2.6
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
libcst==1.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
mongoengine==0.29.1
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
polygon==1.2.6
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-mock==3.14.1
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.1
PyYAML-ft==8.0.0
referencing==0.36.2
requests==2.32.4
respx==0.22.0
rich==14.1.0
rpds-py==0.26.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
strawberry-graphql==0.278.0
strawberry-graphql-django==0.65.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
